backend/mall_api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
pymysql>=1.1.1
